refactor(cloud-fnc): :recycle: Refactored files, daily check
+ daily check is now at 8:00 AM
