+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-  "conventionalCommits.scopes": ["agenda", "icon-picker"]
+  "conventionalCommits.scopes": ["agenda", "icon-picker", "cloud-fnc"]
 }
