+++ functions/src/scheduled/index.ts
@@ -1,0 +1,1 @@
+export * as daily from './daily'
