+++ functions/src/triggers/index.ts
@@ -1,0 +1,1 @@
+export * as operationTriggers from './operations'
